chore: make totals shared
