+++ libs/application/templates/inao/financial-statement-cemetery/src/fields/CemeteryEquities/index.tsx
@@ -13,7 +13,6 @@ import { useLocale } from '@island.is/localization'
 import { FieldBaseProps } from '@island.is/application/types'
 import { InputController } from '@island.is/shared/form-fields'
 import { m } from '../../lib/messages'
-import { Total } from '../KeyNumbers'
 import { getErrorViaPath, getValueViaPath } from '@island.is/application/core'
 import {
   CAPITALNUMBERS,
@@ -24,6 +23,7 @@ import {
 } from '../../utils/constants'
 import { useTotals } from '../../hooks/useTotals'
 import { getTotal } from '../../utils/helpers'
+import { Total } from '../../../../shared/components/Total'
 
 export const CemeteryEquities = ({
   application,
