+++ libs/application/templates/inao/financial-statement-cemetery/src/fields/CemeteryOperation/index.tsx
@@ -8,7 +8,6 @@ import {
 import { Application } from '@island.is/application/types'
 import { useLocale } from '@island.is/localization'
 import { m } from '../../lib/messages'
-import { Total } from '../KeyNumbers'
 import { CemetryIncome } from './CemeteryIncome'
 import { CemeteryExpenses } from './CemeteryExpenses'
 import { CemeteryIncomeLimit } from '../CemeteryIncomeLimit/index'
@@ -17,6 +16,7 @@ import { getValueViaPath } from '@island.is/application/core'
 import { taxInfoQuery } from '../../graphql'
 import { CEMETERYOPERATIONIDS, OPERATINGCOST } from '../../utils/constants'
 import { useTotals } from '../../hooks/useTotals'
+import { Total } from '../../../../shared/components/Total'
 
 export const CemeteryOperation = ({
   application,
